Virtualization Orchestration Layer
An orchestration suite provides a number of services and allows the users access to these in the form of service endpoints.A cloud orchestration suite acts as a transparency layer. It combines all the resources and shows them as one(a cloud) to the users. Users can demand variable resources from the orchestration layer and pay as per the usage.